fix: #20 비동기 처리 완료
