+++ src/main/java/backend/chessmate/global/user/controller/UserController.java
@@ -2,17 +2,13 @@ package backend.chessmate.global.user.controller;
 
 import backend.chessmate.global.auth.config.UserPrincipal;
 import backend.chessmate.global.common.response.SuccessResponse;
-import backend.chessmate.global.user.dto.api.UserGames;
-import backend.chessmate.global.user.dto.api.UserPerf;
-import backend.chessmate.global.user.dto.response.StreaksResponse;
-import backend.chessmate.global.user.dto.response.TierResponse;
+import backend.chessmate.global.user.dto.response.streak.StreaksResponse;
 import backend.chessmate.global.user.dto.response.UserInfoResponse;
 import backend.chessmate.global.user.dto.response.UserPerfResponse;
 import backend.chessmate.global.user.entity.GameType;
 import backend.chessmate.global.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
