+++ src/main/java/backend/chessmate/global/user/dto/response/UserPerfResponse.java
@@ -1,8 +1,12 @@
 package backend.chessmate.global.user.dto.response;
 
+import backend.chessmate.global.user.dto.api.UserPerf;
+import backend.chessmate.global.user.dto.response.tier.TierResult;
+import backend.chessmate.global.user.utils.TierUtil;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import reactor.core.publisher.Mono;
 
 @Builder
 @Data
@@ -29,4 +33,63 @@ public class UserPerfResponse {
     private Integer nowWinningStreak; //현재 연승
     private Integer nowLosingStreak; //현재 연패
 
+    public static UserPerfResponse from(Mono<UserPerf> perf, TierUtil tierUtil) {
+
+        return perf.map(p -> {
+            var stat = p.getStat();
+            Integer highestRating = stat.getHighest().getRating();
+            Integer lowestRating = stat.getLowest().getRating();
+
+            var glicko = p.getPerf().getGlicko();
+            Integer nowRating = (int) glicko.getRating();
+
+            var count = stat.getCount();
+            Integer totalGames = count.getAll();
+            Integer winCount = count.getWin();
+            Integer lossCount = count.getLoss();
+            Integer drawCount = count.getDraw();
+            Long playTime = count.getSeconds();
+
+            var streak = stat.getResultStreak();
+            Integer maxWinningStreak = streak.getWin().getMax().getValue();
+            Integer maxLosingStreak = streak.getLoss().getMax().getValue();
+            Integer nowWinningStreak = streak.getWin().getCur().getValue();
+            Integer nowLosingStreak = streak.getLoss().getCur().getValue();
+
+            Double percentile = p.getPercentile();
+
+            TierResult nowTierResult = tierUtil.calculateTier(nowRating);
+            TierResult maxTierResult = tierUtil.calculateTier(highestRating);
+            TierResult minTierResult = tierUtil.calculateTier(lowestRating);
+
+            double winRate = 0.0;
+            if (totalGames != null && totalGames > 0 && winCount != null) {
+                winRate = (double) winCount / totalGames * 100;
+            }
+
+            return UserPerfResponse.builder()
+                    .nowTier(nowTierResult)
+                    .maxTier(maxTierResult)
+                    .minTier(minTierResult)
+                    .totalGames(totalGames)
+                    .playTime(playTime)
+                    .percentile(percentile)
+                    .winCount(winCount)
+                    .lossCount(lossCount)
+                    .drawCount(drawCount)
+                    .winRate(winRate)
+                    .maxWinningStreak(maxWinningStreak)
+                    .maxLosingStreak(maxLosingStreak)
+                    .nowWinningStreak(nowWinningStreak)
+                    .nowLosingStreak(nowLosingStreak)
+                    .build();
+            
+            
+                }
+                
+        ).block();
+        
+    }
 }
+
+
