+++ src/main/java/backend/chessmate/global/user/service/UserService.java
@@ -10,6 +10,9 @@ import backend.chessmate.global.user.dto.api.UserGame;
 import backend.chessmate.global.user.dto.api.UserGames;
 import backend.chessmate.global.user.dto.response.*;
 import backend.chessmate.global.user.dto.api.UserPerf;
+import backend.chessmate.global.user.dto.response.streak.Streak;
+import backend.chessmate.global.user.dto.response.streak.StreaksResponse;
+import backend.chessmate.global.user.dto.response.tier.TierResult;
 import backend.chessmate.global.user.entity.GameType;
 import backend.chessmate.global.user.utils.LichessUtil;
 import backend.chessmate.global.user.utils.TierUtil;
@@ -17,6 +20,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
 
 import java.time.Instant;
 import java.time.LocalDate;
@@ -33,135 +37,53 @@ public class UserService {
     private final LichessUtil lichessUtil;
     private final TierUtil tierUtil;
 
-//    @Value("${spring.data.redis.key.account_key}")
-//    private String REDIS_ACCOUNT_KEY;
 
-    @Value("${spring.data.redis.key.perf_key}")
-    private String REDIS_PERF_KEY;
 
-    @Value("${spring.data.redis.key.games_key}")
-    private String REDIS_GAMES_KEY;
+    @Value("${spring.data.redis.key.perf_key_base}")
+    private String REDIS_PERF_KEY_BASE;
 
-//    public TierResponse processUserAccount(GameType gameType, UserPrincipal u) {
-//
-//        User user = u.getUser();
-//
-//        UserAccountResponse account;
-//        // 레디스에 UserAccount가 캐싱 되어 있지 않으면 캐싱 후 조회
-//        if (!redisService.hasKey(REDIS_ACCOUNT_KEY + user.getLichessId())) {
-//            account = lichessUtil.getUserAccount(user.getLichessId());
-//        } else {
-//            //이미 캐싱 되어있으므로 레디스 내에서 조회한다.
-//            account = redisService.get(REDIS_ACCOUNT_KEY + user.getLichessId(), UserAccountResponse.class);
-//
-//        }
-//
-//        int rating = getRatingByGameType(account, gameType);
-//        TierResult result = tierUtil.calculateTier(rating);
-//
-//
-//        return TierResponse.builder()
-//                .userName(user.getLichessId())
-//                .gameType(gameType)
-//                .result(result)
-//                .build();
-//
-//    }
-//
-//    private int getRatingByGameType(UserAccountResponse account, GameType gameType) {
-//        return switch (gameType) {
-//            case RAPID -> account.getPerfs().getRapid().getRating();
-//            case BLITZ -> account.getPerfs().getBlitz().getRating();
-//            case BULLET -> account.getPerfs().getBullet().getRating();
-//            case CLASSICAL -> account.getPerfs().getClassical().getRating();
-//            default -> throw new UserException(UserErrorCode.NOT_SUPPORT_GAME_TYPE);
-//        };
-//    }
+    @Value("${spring.data.redis.key.games_key_base}")
+    private String REDIS_GAMES_KEY_BASE;
+
+    @Value("${lichess.ttl.account}")
+    private long acountTTL;
+
+    @Value("${lichess.ttl.perf}")
+    private long perfTTL;
+
 
     public UserPerfResponse processUserPerf(GameType gameType, UserPrincipal u) {
-
         User user = u.getUser();
+        
+        String key = REDIS_PERF_KEY_BASE + ":" + gameType + ":" + user.getLichessId();
 
-        UserPerf perf;
-        if (!redisService.hasKey(REDIS_PERF_KEY + user.getLichessId() + gameType)) {
-            perf = lichessUtil.callUserPerfApi(user, gameType);
-        } else {
-            perf = redisService.get(REDIS_PERF_KEY + user.getLichessId() + gameType, UserPerf.class);
+        Mono<UserPerf> perf = lichessUtil.callUserPerfApi(user, gameType);
+        UserPerfResponse response = UserPerfResponse.from(perf, tierUtil);
 
+        
+        if (redisService.hasKey(REDIS_PERF_KEY_BASE + ":" + gameType + ":" + user.getLichessId())) {
+            log.info("==== Redis 캐시 hit ====");
+            return redisService.get(key, UserPerfResponse.class);
+            
         }
 
-
-        if (perf == null) {
-            log.error("redis perf is null");
-            throw new UserException(UserErrorCode.FAILD_GET_USER_PERF);
-
-        }
-
-        var stat = perf.getStat();
-        Integer highestRating = stat.getHighest().getRating();
-        Integer lowestRating = stat.getLowest().getRating();
-
-        var glicko = perf.getPerf().getGlicko();
-        Integer nowRating = (int) glicko.getRating();
-
-        var count = stat.getCount();
-        Integer totalGames = count.getAll();
-        Integer winCount = count.getWin();
-        Integer lossCount = count.getLoss();
-        Integer drawCount = count.getDraw();
-        Long playTime = count.getSeconds();
-
-        var streak = stat.getResultStreak();
-        Integer maxWinningStreak = streak.getWin().getMax().getValue();
-        Integer maxLosingStreak = streak.getLoss().getMax().getValue();
-        Integer nowWinningStreak = streak.getWin().getCur().getValue();
-        Integer nowLosingStreak = streak.getLoss().getCur().getValue();
-
-        Double percentile = perf.getPercentile();
-
-        // tier 계산
-        TierResult nowTierResult = tierUtil.calculateTier(nowRating);
-        TierResult maxTierResult = tierUtil.calculateTier(highestRating);
-        TierResult minTierResult = tierUtil.calculateTier(lowestRating);
-
-        // 승률 계산
-        double winRate = 0.0;
-        if (totalGames != null && totalGames > 0 && winCount != null) {
-            winRate = (double) winCount / totalGames * 100;
-        }
-
-        return UserPerfResponse.builder()
-                .nowTier(nowTierResult)
-                .maxTier(maxTierResult)
-                .minTier(minTierResult)
-                .totalGames(totalGames)
-                .playTime(playTime)
-                .percentile(percentile)
-                .winCount(winCount)
-                .lossCount(lossCount)
-                .drawCount(drawCount)
-                .winRate(winRate)
-                .maxWinningStreak(maxWinningStreak)
-                .maxLosingStreak(maxLosingStreak)
-                .nowWinningStreak(nowWinningStreak)
-                .nowLosingStreak(nowLosingStreak)
-                .build();
-
-
+        log.info("==== Redis 캐시 miss ==== API 호출 ====");
+        redisService.save(key, response, perfTTL);
+        return response;
     }
 
 
     public GamesInUserInfo processUserGamesInUserInfo(UserPrincipal u) {
-        log.info("=== getUserGames 호출됨 ===");
+        log.info("==== getUserGames 호출됨 ====");
         User user = u.getUser();
 
-        String key = REDIS_GAMES_KEY + user.getLichessId();
+        String key = REDIS_GAMES_KEY_BASE + user.getLichessId();
         UserGames games = null;
         if (redisService.hasKey(key)) {
-            log.info("=== Redis 캐시 hit ===");
+            log.info("==== Redis 캐시 hit ====");
             games = redisService.get(key, UserGames.class);
         } else {
-            log.info("=== Redis 캐시 miss === API 호출 ===");
+            log.info("==== Redis 캐시 miss === API 호출 ====");
             games = lichessUtil.callUserGamesApi(u.getUser()).block();
         }
         Map<String, Long> openingMap = new HashMap<>();
@@ -196,15 +118,13 @@ public class UserService {
                 .opening(opening)
                 .firstMove(firstMove)
                 .build();
-
-
     }
 
     public StreaksResponse processUserGamesInYearStreak(UserPrincipal u) {
         log.info("=== getUserGames 호출됨 ===");
         User user = u.getUser();
 
-        String key = REDIS_GAMES_KEY + user.getLichessId();
+        String key = REDIS_GAMES_KEY_BASE + user.getLichessId();
         UserGames games = null;
         if (redisService.hasKey(key)) {
             log.info("=== Redis 캐시 hit ===");
@@ -268,7 +188,7 @@ public class UserService {
 
         public UserInfoResponse getUserInfo (UserPrincipal u){
             User user = u.getUser();
-            String key = REDIS_GAMES_KEY + user.getLichessId();
+            String key = REDIS_GAMES_KEY_BASE + user.getLichessId();
 
             GamesInUserInfo gamesInUserInfo = processUserGamesInUserInfo(u);
